Main: added single storage for splits, Test:added name-based load handling and also finished building metrics class: need to include more metrics, Train: also model handling, General: added path_builder, serial_data modified to avoid overwrite and serialized and deserialzed accept base file name assuming .pkl later

+++ src/general_utils/general.py
@@ -17,33 +17,61 @@ class General():
                 path(str): The expected path
 
             Returns:
-                None
+                bool: True if exists and None if it is created
+
         """
+
         if not os.path.exists(path):
             os.makedirs(path)
         else:
             print(f"The dir ({path}) exists")
+            return True
+        
+    
+    @staticmethod
+    def path_builder(base:str, name:str)->str:
+        """     
+            Function aimed to build a full path
+
+            Params:
+                base(str): The base path
+                name(str): The filename
+
+            Returns:
+                final_str(str): Composed address
+        """
+
+        return os.path.join(base, name)
+
 
     @staticmethod
-    def serialize_data(data:object, path:str)->None:
+    def serialize_data(data:object, path:str, name:str)->None:
         """ 
             Function aimed to serialize data 
 
             Params:
                 data(object): The data object to store
-                path(str): The  path to store data
-
+                path(str): The base path to store data
+                name(str): The specific filename
             Returns:
                 None
 
         """
-
+        # Check extension
+        if len(name.split('.'))==1:
+            name+='.pkl'
+        
         # Create full path
-        full_path = path + 'train_data'
+        full_path = General.path_builder(path, name )
 
-        # Serialize using pickle
-        with open(full_path, mode = 'wb') as file:
-            pickle.dump(data, file)
+        # Ensure dir existance
+        General.ensure_dir(path)
+
+        #Make sure the file is not already there
+        if not os.path.exists(full_path):
+            # Serialize using pickle
+            with open(full_path, mode = 'wb') as file:
+                pickle.dump(data, file)
         
 
 
@@ -55,14 +83,17 @@ class General():
 
             Params:
                 path(str): The path where data is stored
-                name(str): The name of the specific file where it is
+                name(str): The name of the specific file where data is
 
             Returns:
                 data(DataLoader): The data object 
         """
+        # Check extension
+        if len(name.split('.'))==1:
+            name+='.pkl'
 
         # Create the full path
-        full_path = path + name
+        full_path = General.path_builder(path, name)
 
         # Get data from path
         try:
